deletes actions.ts and moves action interfaces to action-types.ts
