+++ momentum/src/state/ducks/example/reducers.ts
@@ -1,6 +1,7 @@
 import { combineReducers } from 'redux';
-import { TodoTypes } from '../../action-types';
-import { Action } from './actions';
+import { IAddTodoAction, IToggleTodoAction, TodoTypes } from '../../action-types';
+
+type Action = IAddTodoAction | IToggleTodoAction;
 
 // this method can be confusing because it serves two purposes:
 // 1 - it create a new todo
@@ -22,8 +23,6 @@ const todo = (state: {completed: boolean, id: number} | undefined, action: Actio
         ...state,
         completed: !state!.completed
       };
-    default:
-      return state;
   }
 };
 
@@ -33,8 +32,6 @@ const todos = (state = [], action: Action) => {
       return [...state, todo(undefined, action)];
     case TodoTypes.TOGGLE_TODO:
       return state.map(t => todo(t, action));
-    default:
-      return state;
   }
 };
